requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.3.4
python-dotenv==1.0.1

# DeepAgents + Sandboxes
deepagents==0.1.0
langchain-anthropic
anthropic

# OpenSandbox
opensandbox-server
opensandbox-code-interpreter

# 认证与安全
python-jose[cryptography]
passlib[bcrypt]
bcrypt

# 数据库
aiosqlite
sqlalchemy[asyncio]

# WebSocket
websockets

# HTTP 客户端
httpx

# 限流
slowapi

# 测试
pytest
pytest-asyncio
httpx

